+++ src/main/kotlin/com/grepp/nbe1_3_team9/controller/finance/dto/accountBook/AccountBookOneResp.kt
@@ -1,4 +1,4 @@
-package com.grepp.nbe1_3_team9.controller.finance.dto
+package com.grepp.nbe1_3_team9.controller.finance.dto.accountBook
 
 import com.grepp.nbe1_3_team9.domain.entity.finance.Expense
 import java.math.BigDecimal
