+++ src/main/kotlin/com/grepp/nbe1_3_team9/controller/finance/dto/accountBook/AccountBookReq.kt
@@ -1,38 +1,39 @@
-package com.grepp.nbe1_3_team9.controller.finance.dto
+package com.grepp.nbe1_3_team9.controller.finance.dto.accountBook
 
 import com.grepp.nbe1_3_team9.domain.entity.finance.Expense
 import java.math.BigDecimal
 import java.time.LocalDateTime
 
-data class UpdateAccountBookReq(
-    var expenseId: Long = 0L,
+data class AccountBookReq(
     var expenseDate: LocalDateTime = LocalDateTime.now(),
-    var itemName: String = "",
-    var amount: BigDecimal = BigDecimal.ZERO,
-    var paidByUserId: String? = null,
-    var receiptImage: String? = null,
+    val itemName: String = "",
+    val amount: String? = null,
+    val paidByUserId: String? = null,
+    val receiptImage: String? = null,
     var receiptImageByte: ByteArray? = null
 ) {
     companion object {
-        fun toEntity(accountBookReq: UpdateAccountBookReq): Expense {
+        fun toEntity(accountBookReq: AccountBookReq): Expense {
             val expense = Expense()
-            expense.expenseId = accountBookReq.expenseId
             expense.expenseDate = accountBookReq.expenseDate
             expense.itemName = accountBookReq.itemName
-            expense.amount = accountBookReq.amount
+            expense.amount = if (!accountBookReq.amount.isNullOrEmpty()) {
+                BigDecimal(accountBookReq.amount)
+            } else {
+                BigDecimal.ZERO
+            }
             expense.paidBy = accountBookReq.paidByUserId ?: ""
             expense.receiptImage = accountBookReq.receiptImageByte
             return expense
         }
 
-        fun toDTO(expense: Expense): UpdateAccountBookReq {
-            return UpdateAccountBookReq(
-                expenseId = expense.expenseId,
+        fun toDTO(expense: Expense): AccountBookReq {
+            return AccountBookReq(
                 expenseDate = expense.expenseDate,
                 itemName = expense.itemName,
-                amount = expense.amount,
+                amount = expense.amount.toString(),
                 paidByUserId = expense.paidBy,
-                receiptImageByte = expense.receiptImage ?: ByteArray(0)
+                receiptImageByte = expense.receiptImage
             )
         }
     }
