+++ src/main/kotlin/com/grepp/nbe1_3_team9/controller/finance/dto/accountBook/ReceiptDTO.kt
@@ -1,4 +1,4 @@
-package com.grepp.nbe1_3_team9.controller.finance.dto
+package com.grepp.nbe1_3_team9.controller.finance.dto.accountBook
 
 data class ReceiptDTO (
     val expenseDate: String?,
