+++ src/main/kotlin/com/grepp/nbe1_3_team9/controller/finance/dto/exchangeRate/ExchangeRateReqDTO.kt
@@ -1,4 +1,4 @@
-package com.grepp.nbe1_3_team9.controller.finance.dto
+package com.grepp.nbe1_3_team9.controller.finance.dto.exchangeRate
 
 data class ExchangeRateReqDTO(
     val toCountry: String = "",
