+++ src/main/kotlin/com/grepp/nbe1_3_team9/controller/finance/dto/financialPlan/FinancialPlanDTO.kt
@@ -1,4 +1,4 @@
-package com.grepp.nbe1_3_team9.controller.finance.dto
+package com.grepp.nbe1_3_team9.controller.finance.dto.financialPlan
 
 import com.grepp.nbe1_3_team9.domain.entity.finance.FinancialPlan
 import java.math.BigDecimal
