+++ src/main/kotlin/com/grepp/nbe1_3_team9/domain/service/finance/ExchangeRateService.kt
@@ -3,8 +3,8 @@ package com.grepp.nbe1_3_team9.domain.service.finance
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.grepp.nbe1_3_team9.common.exception.ExceptionMessage
 import com.grepp.nbe1_3_team9.common.exception.exceptions.ExchangeRateException
-import com.grepp.nbe1_3_team9.controller.finance.dto.ExchangeRateReqDTO
-import com.grepp.nbe1_3_team9.controller.finance.dto.ExchangeRateResDTO
+import com.grepp.nbe1_3_team9.controller.finance.dto.exchangeRate.ExchangeRateReqDTO
+import com.grepp.nbe1_3_team9.controller.finance.dto.exchangeRate.ExchangeRateResDTO
 import com.grepp.nbe1_3_team9.domain.entity.finance.ExchangeRate
 import com.grepp.nbe1_3_team9.domain.repository.finance.ExchangeRateRepository
 import jakarta.persistence.EntityManager
